pyproject.toml
[tool.poetry]
name = "refferal"
version = "0.1.0"
description = ""
authors = ["Artem Surzhikov <surzhikov64@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.28"
aiosqlite = "^0.20.0"
uvicorn = {extras = ["standart"], version = "^0.28.0"}
pydantic = "^2.6.4"
python-multipart = "^0.0.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "4.0.1"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
httpx = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
